bot/requirements.txt
requests~=2.32.3
numpy~=1.26.4
faiss-cpu~=1.8.0.post1
sentence-transformers~=3.1.1
pyTelegramBotAPI~=4.22.1
loguru~=0.7.2